Adding Authentication/Ownership

+++ routes/listingRoutes.js
@@ -27,6 +27,7 @@ router.post('/',isLoggedin, async (req,res,next)=>{
     if(!newListing){
       next(new ExpressError(400,"Listing doesn't exist !!"));
     }
+    newListing.owner = req.user._id;
     await newListing.save();
     req.flash('success', "New Listing created successfully !");
     res.redirect('/listings');
@@ -42,7 +43,12 @@ router.post('/',isLoggedin, async (req,res,next)=>{
 router.get('/:id', async (req,res,next) =>{
   try{
     let {id} = req.params;
-    const listing = await Listing.findById(id).populate('reviews');
+    const listing = await Listing.findById(id).populate({
+      path: 'reviews',
+      populate:{
+        path: 'author'
+      }
+    }).populate('owner');
     if(!listing){
       return next(new ExpressError(400,"Listing invalid !!"));
     }
@@ -61,11 +67,16 @@ router.get('/:id/edit', isLoggedin, async (req,res,next)=>{
   }catch(err){
     next(err);
   }
-})
+});
 router.put('/:id',isLoggedin, async (req,res,next)=>{
   try{
     if(!req.params.id || !req.body.listing){
-      next(new ExpressError(400,"Give a proper listing"));
+      return next(new ExpressError(400,"Give a proper listing"));
+    }
+    let currListing = await Listing.findById(req.params.id);
+    if(!req.user._id.equals(currListing.owner._id)){
+      req.flash('failure', 'You dont have the permission to Edit!');
+      return res.redirect('/listings');
     }
     await Listing.findByIdAndUpdate(req.params.id, req.body.listing);
     res.redirect(`/listings/${req.params.id}`);
@@ -73,8 +84,14 @@ router.put('/:id',isLoggedin, async (req,res,next)=>{
     next(err);
   }
 })
-router.delete('/:id', isLoggedin, async (req,res) =>{
+router.delete('/:id', isLoggedin, async (req,res,next) =>{
   try{
+    let currListing = await Listing.findById(req.params.id);
+    if(!req.user._id.equals(currListing.owner._id)){
+      req.flash('failure', 'You dont have the permission to Delete!');
+      return res.redirect('/listings');
+    }
+
     await Listing.findByIdAndDelete(req.params.id);
     req.flash('success','Listing deleted successfully!');
     res.redirect('/listings');
