+++ routes/reviewRoutes.js
@@ -2,13 +2,13 @@ const express = require('express');
 const router = express.Router({mergeParams:true});
 const Listing = require('../models/listing');
 const Review = require('../models/review');
-const ExpressError = require('../util/ExpressError');
 
 
 router.post('/', async (req,res,next)=>{
   try{
     let listing = await Listing.findById(req.params.id);
     let newReview = new Review(req.body.review);
+    newReview.author = req.user._id;
     listing.reviews.push(newReview);
 
     await newReview.save();
@@ -23,6 +23,12 @@ router.post('/', async (req,res,next)=>{
 router.delete('/:reviewId', async (req,res,next) =>{
   try{
     let {id,reviewId} = req.params;
+    let currReview = await Review.findById(reviewId);
+    if(currReview.author != req.user._id){
+      req.flash('failure', 'You cannot delete this review!');
+      return res.redirect(`/listings/${id}`);
+    }
+
     await Listing.findByIdAndUpdate(id,{$pull : {reviews:reviewId}});
     await Review.findByIdAndDelete(reviewId);
     req.flash('success', 'Review Deleted');
